+++ tests/testthat/test-reweightView.R
@@ -48,15 +48,8 @@ test_that("reweightView works", {
   expect_equal(curr_new_wts2[1,3], 2)
   expect_equal(curr_new_wts2[-1,], curr_new_wts[-1,])
   
-  expect_equal(app$get_js("$('[data-ns-prefix=reweightInfo-]').css('display')"), "none")
-  app$set_inputs(`reweightInfo-metric_name` = "covr_coverage")
-  expect_equal(app$get_js("$('[data-ns-prefix=reweightInfo-]').css('display')"), "block")
-  
-  app$set_inputs(`reweightInfo-metric_weight` = 2)
-  app$wait_for_idle()
-  expect_equal(app$get_value(input = "reweightInfo-metric_weight"), 0)
-  
   app$set_inputs(`reweightInfo-metric_name` = curr_new_wts[3,1])
+  app$wait_for_idle()
   app$set_inputs(`reweightInfo-metric_weight` = 3.5)
   app$wait_for_idle()
   app$click("reweightInfo-update_weight")
@@ -88,8 +81,8 @@ test_that("reweightView works", {
   expect_equal(dbSelect("select * from package", db_backup)[["decision_id"]], 1)
   
   app$click("reweightInfo-update_pkg_risk")
-  app$click("reweightInfo-confirm_update_risk")
   app$wait_for_idle()
+  app$click("reweightInfo-confirm_update_risk")
   
   db_backup <- app$get_download("reweightInfo-download_database_btn")
   app$wait_for_idle()
@@ -103,4 +96,18 @@ test_that("reweightView works", {
   curr_new_wts <- app$get_value(export = "reweightInfo-curr_new_wts")
   expect_equal(metric_weights, curr_new_wts[,1:2])
   expect_equal(metric_weights[1:3,2], c(2,1,3.5))
+  
+  # Verify that app will not crash and metrics update if database is empty
+  dbUpdate("DELETE FROM package", app_db_loc)
+  
+  expect_equal(nrow(dbSelect("SELECT DISTINCT name FROM package", app_db_loc)), 0)
+  expect_equal(app$get_value(input = "reweightInfo-metric_name"), curr_new_wts[1,1])
+  app$set_inputs(`reweightInfo-metric_weight` = 1)
+  app$click("reweightInfo-update_weight")
+  expect_equal(app$get_value(export = "metric_weights")[1,2], 2)
+  
+  app$click("reweightInfo-update_pkg_risk")
+  app$wait_for_idle()
+  app$click("reweightInfo-confirm_update_risk")
+  expect_equal(app$get_value(export = "metric_weights")[1,2], 1)
 })